# CudnnFind() usage improvements