@DickJC123 2018-10-12
I'm still resolving issues with the CI test I introduced.  The rest of the PR is complete, but not yet pushed.