@DickJC123 2018-10-14
@KellenSunderland @piiswrong @eric-haibin-lin @zheng-da

It took a bit of playing with my newly introduced 'test_large_models' test to get it to fail on the correct CI nodes and not others.  The logs of commit 7bf8d51show the proper misbehavior present with the current codebase.  I've now committed the PR code, which demonstrates the issue is fixed.  See initial description for additional PR benefits.

As I look over the diffs now, I see the PR might look larger than it really is.  Much of the PR change involves moving most of SelectAlgo() code intact into a new callback function CuDNNAlgoSetter.

Also, I see now that I didn't mention introducing an environment variable MXNET_GPU_MEM_LARGE_ALLOC_ROUND_SIZE that defaults to the ideal value 2MB- my preference here rather than hardcoding a magic constant.