@DickJC123 2018-10-14
To remotely debug the issues with my new test as run on the CI systems, I added stderr output that will be emitted on the same nosetests output line (even without the '-s').  Here shown during a failing run:

test_gluon_gpu.test_large_models ...  testing global memory of size 7988903936 ...  1200x1200  1190x1190  1180x1180  1170x1170 
ERROR