@KellenSunderland 2018-10-22
Did some testing over the weekend and can verify this fixes a number of issues I've seen including our recent batch/slice test failures.  Will do a quick syntax pass tomorrow.