@roywei 2018-10-23
@eric-haibin-lin @szha @zheng-da @anirudh2290 @sandeep-krishnamurthy @nswamy   could you help review? Thanks a lot!
This could fix #https://github.com/apache/incubator-mxnet/issues/12767 which is failing in multiple PRs