@KellenSunderland 2018-10-23
Agree with @roywei, this PR will fix many issues and it would be great to get it merged soon.  Let's still do due diligence in terms of review as it's an important component.