@eric-haibin-lin 2018-10-25
@DickJC123 could you address the build issue in CI?
```
src/c_api/c_api.cc:130:43: error: 'free_mem64' may be used uninitialized in this function [-Werror=maybe-uninitialized]

   *free_mem = static_cast<int>(free_mem64);

                                           ^

src/c_api/c_api.cc:131:45: error: 'total_mem64' may be used uninitialized in this function [-Werror=maybe-uninitialized]

   *total_mem = static_cast<int>(total_mem64);

```