@DickJC123 2018-10-26
@eric-haibin-lin I've fixed the latest uninitialized var issue you pointed out and am passing CI.  Thanks to all for the careful review! @KellenSunderland has been following even the last tweeks and approves this PR.