@KellenSunderland 2018-10-26
Yep, updates look good, did some testing and it fixes a few issues (including some CI failures we've been getting recently).

Thanks for the quick iteration @DickJC123.